Add check if the theme "Tags" is an array
The function `$theme->get( 'Tags' )` may return `false` instead of an
array. This can happen, if the current theme was deleted or could not
be loaded correctly. In such cases, WordPress will fall back to another
installed theme. But the missing check is causing an error before that
fallback happens, causing the site in a broken state, that is harder to
recover from.